requirements.txt
flask==2.3.3
sqlalchemy==2.0.23
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
python-dotenv==1.0.0
wtforms==3.1.0
flask-wtf==1.2.1
pandas==2.1.3
plotly==5.17.0
gunicorn==21.2.0
psycopg2-binary==2.9.9